refactor: Popover ui 비즈니스 관심사 분리 [TM-131]
