+++ src/features/calendar/lib/index.ts
@@ -1,2 +1,3 @@
-export * from './getCalendarHeight';
+export * from './calculatePopoverPosition';
+export * from './usePopoverAnimation';
 export * from './usePopoverRender';
